Corre

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -11,7 +11,7 @@
  */
 size_t print_list(const list_t *h)
 {
-	int i = 0;
+	size_t i = 0;
 
 	while (h != NULL)
 	{
@@ -21,7 +21,7 @@ size_t print_list(const list_t *h)
 		}
 		else
 		{
-		printf("[%lu] %s\n", h->len, h->str);
+		printf("[%u] %s\n", h->len, h->str);
 		}
 		h = h->next;
 		i++;
